Rename clipboards module to clipboard.
