+++ clipmanager/clipboard.py
